+++ src/main/java/com/projet/project/service/AdminService.java
@@ -5,7 +5,7 @@ import com.projet.project.entities.SuperAdmin;
 
 public interface AdminService {
     public void addFournisseur(Fournisseur fournisseur);
-    void supprimerFournisseur(Long id);
+    void supprimerFournisseur(int id);
     public void updateFrounisseur(Fournisseur fournisseur);
     public void afficherListe();
